Move implementation from package elisp into package elisp/private/tools.
See
https://docs.google.com/document/d/1L1JFgjpZ7SrBinb24DC_5nTIELeYDacikcme-YcA7xs/view?tab=t.0#heading=h.cy0m49ne5key.
