+++ elisp/private/tools/load.py
@@ -1,4 +1,4 @@
-# Copyright 2021, 2022, 2024 Google LLC
+# Copyright 2021, 2022, 2024, 2025 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ on it in any way outside the rule implementation."""
 from collections.abc import Iterable
 import pathlib
 
-from elisp import runfiles
+from elisp.private.tools import runfiles
 
 def add_path(run_files: runfiles.Runfiles, args: list[str],
              load_path: Iterable[pathlib.PurePosixPath],
