+++ elisp/private/tools/manifest.py
