+++ elisp/private/tools/module.lds
