+++ elisp/private/tools/platform.h
