+++ elisp/private/tools/run_binary.py
@@ -1,4 +1,4 @@
-# Copyright 2021, 2022, 2023, 2024 Google LLC
+# Copyright 2021, 2022, 2023, 2024, 2025 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -26,9 +26,9 @@ import subprocess
 import sys
 from typing import Optional
 
-from elisp import load
-from elisp import manifest
-from elisp import runfiles
+from elisp.private.tools import load
+from elisp.private.tools import manifest
+from elisp.private.tools import runfiles
 
 def main() -> None:
     """Main function."""
