+++ elisp/proto/BUILD
@@ -197,7 +197,7 @@ py_test(
     ],
     data = [":cat"],
     deps = [
-        "//elisp:runfiles",
+        "//elisp/private/tools:runfiles",
         "@abseil-py//absl/flags",
         "@abseil-py//absl/testing:absltest",
     ],
