+++ emacs/BUILD
@@ -44,7 +44,7 @@ py_test(
     ],
     data = [":emacs"],
     deps = [
-        "//elisp:runfiles",
+        "//elisp/private/tools:runfiles",
         "@abseil-py//absl/flags",
         "@abseil-py//absl/testing:absltest",
     ],
